feat: Add Modal Component with useOpenAndCloseModal Hook
The commit introduces a new `Modal` component along with a custom React Hook, `useOpenAndCloseModal`, to manage the opening and closing of the modal smoothly. The hook helps to avoid issues with modal animations by setting the initial `open` state to `false` and updating it to `true` after the component is fully mounted.

A README file included provides usage, installation, and detailed information about the Modal component and the custom hook, which can help in managing modals effectively in the project. Besides the code changes, the implementation has also been documented with the addition of the interface for typescript props.

This change is crucial to provide a pre-configured way to handle modals' opening and closing animations, reducing the redundancy in the code whenever modals are needed within the application.